END OF THIS CACHE !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

+++ main.py
@@ -65,14 +65,15 @@ def getAllTasks(update = False):
     return tasks
 
 
-def getAllComments(update = False):
-    key = 'comments'
+def getCommentsOfTasks(task_id, update = False):
+    key = task_id
     comments = memcache.get(key)
     if comments is None or update:
-        comments = db.GqlQuery('Select * from Comment ORDER BY task_id, date DESC')        
+        comments = db.GqlQuery('Select * from Comment where task_id = :1 ORDER BY date DESC', int(key) )       
         comments = list(comments)
         memcache.set(key, comments)
-    return comments
+        return comments, True
+    return comments, False
 
 def getTaskByKey(task_id):
     key = db.Key.from_path('Task',int(task_id), parent = task_key())
@@ -85,13 +86,36 @@ def getTaskByKey(task_id):
 ####        DB ACCESS   -   TASK PROCESS
 ##########################################################################################
 
+def putTaskInCache(task):
+    key = 'tasks'
+    tasks = memcache.get(key)
+    if tasks :
+        tasks.append(task)
+    else:
+        tasks = []
+        tasks.append(task)
+
+    memcache.delete(key)
+    memcache.set(key, tasks)        
+
 def participateToTask(task_id, username):    
-    task = getTaskByKey(task_id)    
+    key = db.Key.from_path('Task',int(task_id), parent = task_key())
+    task = db.get(key)   
     #If the task exists, add the user as a participant
-    if task is not None:
-        oldTask = task                               
+    if task is not None:        
         task.addParticipant(int(task_id), username)
-        updateTaskInCache(oldTask, task)
+        # update the cache
+        key = 'tasks'
+        tasks = memcache.get(key)
+        for c in tasks:
+            if c.__eq__(task):
+                index = tasks.index(c)
+                tasks[index] = task
+                break
+        if not memcache.set(key, tasks):            
+            memcache.delete(key)
+            memcache.set(key, tasks)
+        
         return True
     else:
         return False
@@ -102,47 +126,64 @@ def cancelParticipation(task_id, username):
         return False
     participants = task.getParticipant(task_id)    
     if participants is not None and username in participants:
-        oldTask = task
-        task.participants.remove(self.user.username)                
+        task.participants.remove(username)
+        # update the cache
+        key = 'tasks'
+        tasks = memcache.get(key)
+        for c in tasks:
+            if c.__eq__(task):
+                index = tasks.index(c)
+                tasks[index] = task                               
+                break
+        if not memcache.set(key, tasks):            
+            memcache.delete(key)
+            memcache.set(key, tasks)                  
         task.put()
-        updateTaskInCache(oldTask, task)       
+          
         return True
     return False
 
 
 def deleteTask(task_id, username):
     if task_id and username:         
-        task = getTaskByKey(task_id)    
-        #Get the username of the creator 
+        key = db.Key.from_path('Task',int(task_id), parent = task_key())
+        task = db.get(key)  
+        #Get the username of the creator
         if task is not None:                
             creator = task.author;
             if username == creator:
-                task_deleted = task
-                task.deleteComments(int(task_id))
+                #Delete entry in cache
+                key = 'tasks'
+                tasks = memcache.get(key)
+
+                found = False
+                for c in tasks:
+                    if c.__eq__(task):
+                        tasks.remove(task)
+                        found = True
+                        break
+                if not found :
+                    return list('notFound !!!'), False
+                
+                memcache.delete(key)
+                if not memcache.set(key, tasks):            
+                    memcache.delete(key)
+                    memcache.set(key, tasks)
+                
                 task.delete()
-                deleteTaskInCache(task_deleted)                
-                return (True, "Task has been correctly deleted")
+                #Delete all comments relative to the Task        
+                taskComments = Comment.all()
+                taskComments.filter("task_id =", task_id)
+                for comment in taskComments.run():
+                    comment.delete()
+
+                return memcache.get(key), True
             else:
-                return (False, "You cannot delete tasks from other Bros modafucka !!!")
+                return list('notasks'), False
     else:
-        return (False, "Task has not been deleted : data are missing")
+        return list('notasks'), False
 
-def updateTaskInCache(oldTask, newTask):
-    key = 'tasks'
-    tasks = memcache.get(key)
-    for t, i in tasks, range(tasks):
-        if t.__eq__(oldTask):
-            tasks[i] = newTask
-            break
-    memcache.set(key, tasks)
 
-def deleteTaskInCache(task):
-    key = 'tasks'
-    tasks = memcache.get(key)
-    for t, i in tasks, range(tasks):
-        if t.__eq__(task):
-            tasks.pop(i)
-    memcache.set(key, tasks)
 
 
 ##########################################################################################
@@ -152,20 +193,51 @@ def deleteTaskInCache(task):
 def commentTask(task_id, author, content):
     if task_id and content :
         com = Comment(task_id = int(task_id), author = author, content = content, date = datetime.datetime.now())
-        com.put()
-        getAllComments(True)
-        return True
+        com.put()      
+        # Update the cache
+        if com :
+            comments = memcache.get(task_id)
+            if comments :
+                comments.append(com)
+                memcache.set(task_id, comments)
+            else:
+                comments = []
+                comments.append(com)
+                memcache.set(task_id, comments)
+
+            
+            return True
+        else:
+            return False
     else:
         return False
 
-def deleteComment(comment_id):
+def deleteComment(task_id, comment_id):
     key_comment = db.Key.from_path('Comment', int(comment_id))
     commentToDel = db.get(key_comment)
     if commentToDel:
+        #Delete entry in the cache
+        comments = memcache.get(task_id)
+        if len(comments) == 0:
+            pass
+        elif len(comments) == 1:
+            comments.pop(0)
+        else:
+            for c in comments:
+                if c.__eq__(commentToDel):
+                    index = comments.index(c)
+                    comments.pop(index)            
+                    break
+        
+        memcache.set(task_id, comments) 
         commentToDel.delete()
-        getAllComments(True)
-        return True
-    return False
+         
+
+        return memcache.get(task_id), True
+    return list('nocomment'), False
+
+
+
 
 
 class HurbHandler(webapp2.RequestHandler):
@@ -204,13 +276,16 @@ class HurbHandler(webapp2.RequestHandler):
         
 
 class Main(HurbHandler):
+    # def get(self, tasks=None):
+    #     if tasks is None:
+    #         tasks = getAllTasks(True)
     def get(self):
         tasks = getAllTasks()
-        
+                   
         if self.user:            
             self.render('home.html', tasks = tasks, username = self.user.username)            
         else:            
-            self.render('home.html', tasks = tasks)
+            self.render('home.html', tasks = tasks, username = "")
 
     def post(self):
         
@@ -221,40 +296,40 @@ class Main(HurbHandler):
         self.participate = self.request.get('participateTask')
         self.cancel = self.request.get('cancelParticipation')
 
+        error = False
+        error_log = ""
+
         #comment the Task
         if self.comment :
-            if not commentTask(self.task_id, self.user.username, self.content):                
-                self.response.out.write("Comment has not been saved")
-            else:                                
-                self.redirect('/')
-
-        #Delete the task with its comments
+            if not commentTask(self.task_id, self.user.username, self.content):   
+                error = True             
+                error_log = "Comment has not been saved"
+            
+        #Delete the task with its comments        
         if self.delete :
-            status,msg = deleteTask(self.task_id, self.user.username)
-            if not status:                
-                self.response.out.write("%s" % msg)
-            else:
-                self.response.out.write("%s" % msg)
+            tasks, status = deleteTask(self.task_id, self.user.username)
+            if status:
                 self.redirect('/')
+            else:
+                self.redirect('/', None)
 
         #participate to the task
-        if self.participate :
-            #self.response.out.write("You really want to participate !!!")
+        if self.participate :            
             if not participateToTask(self.task_id, self.user.username):                
-                self.response.out.write("Task doesn't exist anymore")
-            else:
-                self.redirect('/')
-
+                error = True
+                error_log = "Task doesn't exist anymore"
+        
         #Leave a task:
         if self.cancel: 
-            if cancelParticipation(self.task_id, self.user.username):
-                self.redirect('/')
+            if not cancelParticipation(self.task_id, self.user.username):
+                error = True
+                error_log = "Participation canceled not handled"
 
-
-        tasks = getAllTasks()        
-        comments = getAllComments()
-        self.render('home.html', tasks = tasks, comments = comments)
-
+                
+        if error :
+            self.response.out.write("error log : "+error_log)
+        else:            
+            self.redirect('/')            
 
 
 class NewTask(HurbHandler):
@@ -309,7 +384,7 @@ class NewTask(HurbHandler):
             self.render('newtask.html', **params)
         else:
             task = Task(parent = task_key(), author = self.author, title = self.taskTitle, description = self.taskDescription, date = convertedDate, location_lat = self.location_lat, location_lng = self.location_lng, reward = self.reward, participants = [])           
-            task.put()         
+            task.put()                   
             self.redirect('/task/%s' % str(task.key().id())) 
 
 
@@ -318,13 +393,16 @@ class TaskPage(HurbHandler):
         task = getTaskByKey(task_id)
         if not task:
             self.error(404)
-            return 
+            return
         tasks = getAllTasks(True)        
-        comments = getAllComments()  
+        
+        comments, status = getCommentsOfTasks(task_id)
+
         if self.user:
             self.render("taskpermalink.html", task=task, username = self.user.username, comments = comments)
         else:
-            self.render("taskpermalink.html", task=task, username = "", comments = comments)
+            self.response.out.write("GET : no user")            
+            
 
     def post(self, task_id):
         url_get = self.request.url
@@ -341,47 +419,64 @@ class TaskPage(HurbHandler):
         self.deleteComment = self.request.get('deleteCom')
         self.comment_id = self.request.get('comment_id')
 
+        error_log = ""
+        error = False
+
         task = getTaskByKey(task_id)
         if not task:
-            self.error(404)
+            error_log = "task not found"
+            error = True
+            #self.error(404)
             return
+
+        
 
         #comment the Task
         if self.comment :
-            if not commentTask(self.task_id, self.user.username, self.content):                
-                self.response.out.write("Comment has not been saved")
-            else:
-                self.redirect(redirectTo)
+            if not commentTask(task_id, self.user.username, self.content):                
+                error_log = "Comment has not been saved"
+                error = True
 
-        #Delete the task with its comments
+
+        #Delete the task with its comments        
         if self.delete :
-            (status,msg) = deleteTask(self.task_id, self.user.username)
-            if not status:                
-                self.response.out.write("%s" % msg)
+            tasks, status = deleteTask(self.task_id, self.user.username)            
+            if status:
+                redirectTo = "/"
+                # error_log = "delete task correct "
+                # for t in tasks :
+                #     error_log = error_log + t.title + " - "
+                # error = True
+                #self.redirect('/', tasks)
+                #self.redirect('/')
             else:
-                self.response.out.write("%s" % msg)
-                self.redirect(redirectTo)
+                error_log = "delete task NOT correct "
+                for t in tasks :
+                    error_log = error_log + t + " - "
+                error = True
+                # self.redirect('/', None)
+
 
         if self.deleteComment:
-            if deleteComment(self.comment_id):
-                self.redirect(redirectTo)
+            comments, status = deleteComment(task_id, self.comment_id)             
 
 
         #participate to the task
         if self.participate :            
-            if not participateToTask(self.task_id, self.user.username):                
-                self.response.out.write("Task doesn't exist anymore")
-            else:
-                self.redirect(redirectTo)
-
+            if not participateToTask(task_id, self.user.username):                
+                error_log = "Task doesn't exist anymore"
+                error = True
+            
         #Leave a task:
         if self.cancel: 
-            if cancelParticipation(self.task_id, self.user.username):
-                self.redirect(redirectTo)
+            if not cancelParticipation(task_id, self.user.username):
+                error_log = "Cannot cancel participation"
+                error = True
 
-        comments = getAllComments()  
-        self.render("taskpermalink.html", task=task, username = self.user.username, comments = comments)
-        
+        if error :
+            self.response.out.write("Error : "+error_log)
+        else:
+            self.redirect(redirectTo)
 
 
 class Signup(HurbHandler):
